+++ tests/Classifiers/RandomForestTest.php
@@ -65,7 +65,7 @@ class RandomForestTest extends TestCase
         $this->assertFalse($this->estimator->trained());
     }
 
-    public function test_train_predict()
+    public function test_train_predict_feature_importances()
     {
         $training = $this->generator->generate(self::TRAIN_SIZE);
         
@@ -82,18 +82,11 @@ class RandomForestTest extends TestCase
         $score = $this->metric->score($predictions, $testing->labels());
 
         $this->assertGreaterThanOrEqual(self::MIN_SCORE, $score);
-    }
-
-    public function test_train_feature_importances()
-    {
-        $training = $this->generator->generate(self::TRAIN_SIZE);
-
-        $this->estimator->train($training);
 
         $importances = $this->estimator->featureImportances();
 
         $this->assertCount(3, $importances);
-        $this->assertEquals(1, array_sum($importances));
+        $this->assertEquals(1., array_sum($importances));
     }
 
     public function test_train_with_unlabeled()