+++ tests/Regressors/ExtraTreeRegressorTest.php
@@ -62,7 +62,7 @@ class ExtraTreeRegressorTest extends TestCase
         $this->assertEquals(0, $this->estimator->height());
     }
 
-    public function test_train_predict()
+    public function test_train_predict_feature_importances()
     {
         $this->estimator->train($this->training);
 
@@ -75,16 +75,11 @@ class ExtraTreeRegressorTest extends TestCase
         $score = $this->metric->score($predictions, $this->testing->labels());
 
         $this->assertGreaterThanOrEqual(self::MIN_SCORE, $score);
-    }
-
-    public function test_train_feature_importances()
-    {
-        $this->estimator->train($this->training);
 
         $importances = $this->estimator->featureImportances();
 
         $this->assertCount(6, $importances);
-        $this->assertEquals(1, array_sum($importances));
+        $this->assertEquals(1., array_sum($importances));
     }
 
     public function test_train_with_unlabeled()