+++ tests/Regressors/GradientBoostTest.php
@@ -58,7 +58,7 @@ class GradientBoostTest extends TestCase
         $this->assertFalse($this->estimator->trained());
     }
 
-    public function test_train_predict()
+    public function test_train_predict_feature_importances()
     {
         $training = $this->generator->generate(self::TRAIN_SIZE);
 
@@ -73,6 +73,11 @@ class GradientBoostTest extends TestCase
         $score = $this->metric->score($predictions, $testing->labels());
 
         $this->assertGreaterThanOrEqual(self::MIN_SCORE, $score);
+
+        $importances = $this->estimator->featureImportances();
+
+        $this->assertCount(3, $importances);
+        $this->assertEquals(1., array_sum($importances));
     }
 
     public function test_train_with_unlabeled()